% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demographic_data}
\alias{demographic_data}
\title{Demographic information for study participants.}
\format{
A data frame with 720 rows and 8 variables:
\describe{
\item{ID}{Subject Identification Number - numerical}
\item{Gender}{Subject's gender - 'M', 'F'}
\item{PrimaryDiagnosis}{Individual diagnosis - 'Autism','ADHD','None'}
\item{SecondaryDiagnosis}{Indicates if there is a secondary diagnosis - 'Yes','No'}
\item{ADHD_Subtype}{If the subject meets the criteria for a Primary or Secondary Diagnosis of ADHD,
indicates which subclassification. - ‘Inattentive’, ‘Hyperactive/Impulsive’, ‘Combined’, ‘No dx’}
\item{CurrentlyNotTakingMeds}{Parent report of whether child is currently not taking any medications. - ‘TRUE’, ‘FALSE’}
\item{CurrentlyTakingAtomoxetine}{f CurrentlyNotTakingMeds is ‘FALSE’, make sure not taking Atomoxetine - ‘TRUE’, ‘FALSE’}
\item{CurrentlyTakingClonidine}{If CurrentlyNotTakingMedsis ‘FALSE’, make sure not taking Clonidine - ‘TRUE’, ‘FALSE’}
}
}
\source{
{Amanda Mejia, via Indiana University and Johns Hopkins University.}
}
\usage{
demographic_data
}
\description{
A dataset containing demographic information.
}
\keyword{datasets}
