% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{handedness}
\alias{handedness}
\title{Handedness information for study participants.}
\format{
A data frame with 720 rows and 3 variables:
\describe{
\item{ID}{Subject Identification Number - numerical}
\item{visit}{Indicates first or repeated visits - numerical}
\item{EdinburghHandedness_Integer}{Parents indicated which hand child most likely to use for 12 common
tasks - numerical}
}
}
\source{
{Amanda Mejia, via Indiana University and Johns Hopkins University.}
}
\usage{
handedness
}
\description{
A dataset containing handedness information.
}
\keyword{datasets}
