% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mABC}
\alias{mABC}
\title{Movement Assessment Battery test for children results for study participants.}
\format{
A data frame with 720 rows and 16 variables:
\describe{
\item{ID}{Subject Identification Number - numerical}
\item{visit}{Indicates first or repeated visits - numerical}
\item{mABC_AGE}{Age of the participant at the time of mABC testing - numerical}
\item{mABC_TotalStandardScore}{The Movement Assessment Battery for Children (mABC)
is designed to identify and describe impairments in motor performance of children and adolescents 3 through
17 years of age.The Performance Test involves children completing a series of fine and gross motor tasks
grouped in three categories: Manual Dexterity, Aiming and Catching, and Balance. - numerical}
}
}
\source{
{Amanda Mejia, via Indiana University and Johns Hopkins University.}
}
\usage{
mABC
}
\description{
A dataset containing mABC results.
}
\keyword{datasets}
