% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SRS}
\alias{SRS}
\title{Social Responsiveness Scale questionnaire results for study participants.}
\format{
A data frame with 720 rows and 5 variables:
\describe{
\item{ID}{Subject Identification Number - numerical}
\item{visit}{Indicates first or repeated visits - numerical}
\item{SRS_AGE}{Age of the participant at the time of SRS completion - numerical}
\item{SRS_VERSION}{Version of the Social Responsiveness Scale (SRS) Questionnaire administered to parents - 1 or 2}
\item{SRS_TotalRawScore}{The SRS probes a child’s level of motivation to engage in social interactions,
his/her ability to recognize emotional and interpersonal cues from other people, to interpret those cues
correctly, and to respond to what he/she interprets appropriately. The social deficit severity index range
from 0 to 195 with higher scores indicating more severe social deficits. - numerical}
}
}
\source{
{Amanda Mejia, via Indiana University and Johns Hopkins University.}
}
\usage{
SRS
}
\description{
A dataset containing SRS questionnaire results.
}
\keyword{datasets}
