% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WISC}
\alias{WISC}
\title{Wechsler Intelligence Scale for Children results for study participants.}
\format{
A data frame with 720 rows and 15 variables:
\describe{
\item{ID}{Subject Identification Number - numerical}
\item{visit}{Indicates first or repeated visits - numerical}
\item{WISC_AGE}{Age of the participant at the time of WISC completion - numerical}
\item{WISC_VERSION}{Version of the test administered - 4 or 5}
\item{WISC4_Vocabulary}{Vocabulary Score from version 4 - numerical}
\item{WISC4_Information}{General knowledge questions score from version 4 - numerical}
\item{WISC4_Similarities}{Children are asked to describehow two words are alikefrom version 4 - numerical}
\item{WISC4_BlockDesign}{Children put together red and white blocks in a pattern according to a
displayed model from version 4 - numerical}
\item{WISC4_PictureConcepts}{Children are provided with a series of pictures presented in rows (either two or
three rows) and asked to determine which pictures go together, one from each row, from version 4. - numerical}
\item{WISC4_MatrixReasoning}{Children are shown an array of pictures with one missing square and
asked to select the picture that fits the array from five options, from version 4. - numerical}
\item{WISC5_Vocabulary}{From version 5 - numerical}
\item{WISC5_Similarities}{From version 5 - numerical}
\item{WISC5_BlockDesign}{From version 5 -numerical}
\item{WISC5_MatrixReasoning}{From version 5 - numerical}
\item{WISC5_FigureWeights}{Children view pictures of shapes on a scale (or scales) with one empty side and are
asked to select the choice that keeps the scale balanced; from version 5 - numerical}
}
}
\source{
{Amanda Mejia, via Indiana University and Johns Hopkins University.}
}
\usage{
WISC
}
\description{
A dataset containing WISC results.
}
\keyword{datasets}
