% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WISC4Lookup}
\alias{WISC4Lookup}
\title{WISC4 Lookup Resources}
\format{
A data frame with 110 rows and 2 variables:
\describe{
\item{WISC4.ScaledSum}{Scaled Sum to use as a reference to convert WISC4 to GAI - numerical}
\item{WISC4.GAI}{The corresponding GAI scores for WISC4 scaled sums - numerical}
}
}
\source{
{Amanda Mejia, via Indiana University and Johns Hopkins University.}
}
\usage{
WISC4Lookup
}
\description{
A dataset containing resources to convert WISC4 scores to GAI.
}
\keyword{datasets}
