% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WISC5Lookup}
\alias{WISC5Lookup}
\title{WISC5 Lookup Resources}
\format{
A data frame with 92 rows and 3 variables:
\describe{
\item{WISC5.ScaledSum}{Scaled Sum to use as a reference to convert WISC5 to GAI - numerical}
\item{WISC5.GAI}{The corresponding GAI scores for WISC5 scaled sums - numerical}
}
}
\source{
{Amanda Mejia, via Indiana University and Johns Hopkins University.}
}
\usage{
WISC5Lookup
}
\description{
A dataset containing resources to convert WISC5 scores to GAI.
}
\keyword{datasets}
